+++ packages/aesthetic-adapter-jss/tests/UnifiedAdapter.test.js
@@ -13,6 +13,7 @@ import {
   SYNTAX_FONT_FACE,
   SYNTAX_KEYFRAMES,
   SYNTAX_MEDIA_QUERY,
+  SYNTAX_SUPPORTS,
 } from '../../../tests/mocks';
 
 describe('aesthetic-adapter-jss/UnifiedAdapter', () => {
@@ -68,6 +69,9 @@ describe('aesthetic-adapter-jss/UnifiedAdapter', () => {
         ],
         '@media (max-width: 600px)': {
           padding: '4px 8px',
+        },
+        '@supports (display: flex)': {
+          display: 'flex',
         },
       },
     });
@@ -135,6 +139,20 @@ describe('aesthetic-adapter-jss/UnifiedAdapter', () => {
         },
         '@media (min-width: 300px)': {
           color: 'blue',
+        },
+      },
+    });
+  });
+
+  it('supports supports', () => {
+    expect(instance.convert(SYNTAX_SUPPORTS)).toEqual({
+      sup: {
+        display: 'block',
+        '@supports (display: flex)': {
+          display: 'flex',
+        },
+        '@supports not (display: flex)': {
+          float: 'left',
         },
       },
     });