+++ packages/aesthetic-utils/src/injectMediaQueries.js
@@ -10,12 +10,6 @@ import type { MediaQueries, StyleDeclaration } from '../../types';
 
 export default function injectMediaQueries(properties: StyleDeclaration, queries: MediaQueries) {
   Object.keys(queries).forEach((key) => {
-    let query = key;
-
-    if (query.slice(0, 1) !== '(' || query.slice(-1) !== ')') {
-      query = `(${query})`;
-    }
-
-    properties[`@media ${query}`] = queries[key];
+    properties[`@media ${key}`] = queries[key];
   });
 }